perf(reddit/trending): code-split tab panels via next/dynamic (AGN-455)
